update - restructuring: moved base overlay components into common components location
