+++ packages/editor/src/app/components/overlay/index.ts
@@ -1,0 +1,3 @@
+export * from './backdrop';
+export * from './drawer';
+export * from './modal';
