+++ packages/editor/src/app/pages/workspace/components/overlay/asset-browser/index.ts
