--- packages/editor/src/app/pages/workspace/components/overlay/components/index.ts
@@ -1,9 +1,0 @@
-export * from './backdrop';
-export * from './drawer';
-export * from './modal';
-export * from './glossary';
-export * from './asset-browser';
-export * from './resource';
-export * from './publish';
-export * from './template-browser';
-export * from './prompt-project-name';
