+++ packages/editor/src/app/pages/workspace/components/overlay/index.ts
@@ -1,2 +1,6 @@
-export * from './overlay.types';
-export * from './components';
+export * from './asset-browser';
+export * from './glossary';
+export * from './prompt-project-name';
+export * from './publish';
+export * from './resource';
+export * from './template-browser';
