+++ packages/editor/src/app/pages/workspace/components/overlay/prompt-project-name/_prompt-project-name.scss
