+++ packages/editor/src/app/pages/workspace/components/overlay/prompt-project-name/index.ts
