+++ packages/editor/src/app/pages/workspace/components/overlay/publish/publish.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState, useRef, forwardRef } from 'react';
 import { AnimatePresence } from 'framer-motion';
-import '../../_overlay.scss';
-import { Backdrop, Drawer } from '../';
-import { Projects, Settings } from '../../../../../../models';
-import { hasProp } from '../../../../../../utils';
+import '../_overlay.scss';
+import { Backdrop, Drawer } from '../../../../../components';
+import { Projects, Settings } from '../../../../../models';
+import { hasProp } from '../../../../../utils';
 import { CourseSettings } from './course-settings';
 import { Reporting } from './reporting';
 import { ExportOptions } from './export-options';
