+++ packages/editor/src/app/pages/workspace/components/overlay/template-browser/_template-browser.scss
